LayoutTests/media/modern-media-controls/volume-down-support/volume-down-support.html

<!DOCTYPE html>
<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-utils.js" type="text/javascript"></script>
<body>
<video src="../../content/test.mp4" style="width: 320px; height: 240px;" controls></video>
<div id="shadow"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing clicking on the volume down button correctly sets the volume to 0.");

const media = document.querySelector("video");
const button = document.body.appendChild(document.createElement("button"));

media.addEventListener("webkitfullscreenchange", function() {
    if (media.webkitDisplayingFullscreen) {
        window.requestAnimationFrame(() => {
            window.requestAnimationFrame(() => {
                shouldBe("media.volume", "1");

                media.addEventListener("volumechange", () => {
                    shouldBe("media.volume", "0");
                    debug("");
                    media.remove();
                    button.remove();
                    finishJSTest();
                });

                debug("");
                debug("Pressing on the volume down button");
                pressOnElement(window.internals.shadowRoot(media).lastElementChild.lastElementChild.querySelector("button.volume-down"));
            });
        });
    }
});

media.addEventListener("loadedmetadata", () => {
    if (!("eventSender" in window)) {
        debug("This test is designed to run in DRT");
        return;
    }

    // Click a button so we may enter fullscreen.
    button.addEventListener("click", event => {
        try {
            media.webkitEnterFullscreen();
        } catch(e) {
            debug("Toggling fullscreen failed");
            finishJSTest();
        }
    });

    pressOnElement(button);
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
