LayoutTests/media/modern-media-controls/volume-down-support/volume-down-support-expected.txt

Testing clicking on the volume down button correctly sets the volume to 0.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS media.volume is 1

Pressing on the volume down button
PASS media.volume is 0

PASS successfullyParsed is true

TEST COMPLETE
